+++ server/router.ts
@@ -20,13 +20,14 @@ import { guideRoutes } from "./routes/guide";
 import { foreignRoutes } from "./routes/foreign";
 import { alertsRoutes } from "./routes/alerts";
 import { notifyRoutes } from "./routes/notify";
+import { pushRoutes } from "./routes/push";
 import { monitorRoutes } from "./routes/monitor";
 import { supportRoutes } from "./routes/support";
 import { paymentRoutes } from "./routes/payment";
 import { opsRoutes } from "./routes/ops";
 import { pageRoutes } from "./routes/pages";
 
-const ROUTES: RouteFn[] = [metaRoutes, geoRoutes, assessRoutes, analysisRoutes, cropsRoutes, regionFitRoutes, journalRoutes, accountRoutes, marketRoutes, budgetRoutes, guideRoutes, foreignRoutes, notifyRoutes, alertsRoutes, monitorRoutes, supportRoutes, paymentRoutes, opsRoutes, pageRoutes];
+const ROUTES: RouteFn[] = [metaRoutes, geoRoutes, assessRoutes, analysisRoutes, cropsRoutes, regionFitRoutes, journalRoutes, accountRoutes, marketRoutes, budgetRoutes, guideRoutes, foreignRoutes, notifyRoutes, pushRoutes, alertsRoutes, monitorRoutes, supportRoutes, paymentRoutes, opsRoutes, pageRoutes];
 
 export async function route(ctx: Ctx, req: http.IncomingMessage, res: http.ServerResponse, url: URL): Promise<void> {
   for (const handler of ROUTES) {